+++ src/main/kotlin/com/lykke/matching/engine/daos/wallet/Wallet.kt
@@ -1,5 +1,6 @@
 package com.lykke.matching.engine.daos.wallet
 
+import java.math.BigDecimal
 import java.util.HashMap
 
 class Wallet {
@@ -17,7 +18,7 @@ class Wallet {
         }
     }
 
-    fun setBalance(asset: String, balance: Double) {
+    fun setBalance(asset: String, balance: BigDecimal) {
         val oldBalance = balances[asset]
         if (oldBalance == null) {
             balances[asset] = AssetBalance(clientId, asset, balance)
@@ -26,7 +27,7 @@ class Wallet {
         }
     }
 
-    fun setReservedBalance(asset: String, reservedBalance: Double) {
+    fun setReservedBalance(asset: String, reservedBalance: BigDecimal) {
         val oldBalance = balances[asset]
         if (oldBalance == null) {
             balances[asset] = AssetBalance(clientId, asset, reservedBalance, reservedBalance)
